Move playwrite tests
