+++ web/package.json
@@ -13,8 +13,7 @@
 		"setup": "mkdir -p dist/.well-known/appspecific && cp app/index.html dist/ && cp -r app/public/* dist/ && cp app/globals.css dist/ && cp -r app/.well-known/* dist/.well-known/",
 		"test": "bun run test:unit && bun run test:integration",
 		"test:unit": "bun test --define __API_URL__='\"ws://127.0.0.1:1\"' app/",
-		"test:integration": "bun --env-file=../.env run playwright test tests/",
-		"test:e2e": "bun --env-file=../.env run playwright test e2e/",
+		"test:integration": "bun --env-file=../.env run playwright test",
 		"test:setup": "bunx playwright install --with-deps chromium"
 	},
 	"dependencies": {
