+++ web/test.preload.ts
@@ -1,0 +1,5 @@
+Object.defineProperty(globalThis, "__API_URL__", {
+	value: "ws://127.0.0.1:1",
+	writable: true,
+	configurable: true,
+});
